Fixes migration issues. Adds Double support to AODE.

+++ Aurora/Physics/AuroraOpenDynamicsEngine/AODEPrim.cs
@@ -1568,9 +1568,9 @@ Console.WriteLine(" JointCreateFixed");
                     // 'VEHICLES' are dealt with in ODEDynamics.cs
                     m_vehicle.Step(Body, timestep, _parent_scene, this);
                     d.Vector3 vel = d.BodyGetLinearVel(Body);
-                    _velocity = new Vector3(vel.X, vel.Y, vel.Z);
+                    _velocity = new Vector3((float)vel.X, (float)vel.Y, (float)vel.Z);
                     d.Vector3 pos = d.BodyGetPosition(Body);
-                    _position = new Vector3(pos.X, pos.Y, pos.Z);
+                    _position = new Vector3((float)pos.X, (float)pos.Y, (float)pos.Z);
                     _zeroFlag = false;
                 }
                 else
@@ -1654,9 +1654,9 @@ Console.WriteLine(" JointCreateFixed");
                         d.Vector3 pos = d.BodyGetPosition(Body);
                         _target_velocity =
                             new Vector3(
-                                (m_PIDTarget.X - pos.X) * ((/*PID_G - */m_PIDTau) * timestep),
-                                (m_PIDTarget.Y - pos.Y) * ((/*PID_G - */m_PIDTau) * timestep),
-                                (m_PIDTarget.Z - pos.Z) * ((/*PID_G - */m_PIDTau) * timestep)
+                                (float)(m_PIDTarget.X - pos.X) * ((/*PID_G - */m_PIDTau) * timestep),
+                                (float)(m_PIDTarget.Y - pos.Y) * ((/*PID_G - */m_PIDTau) * timestep),
+                                (float)(m_PIDTarget.Z - pos.Z) * ((/*PID_G - */m_PIDTau) * timestep)
                                 );
 
                         //  if velocity is zero, use position control; otherwise, velocity control
@@ -1688,12 +1688,12 @@ Console.WriteLine(" JointCreateFixed");
                             _zeroFlag = false;
 
                             // We're flying and colliding with something
-                            fx = ((_target_velocity.X) - vel.X) * (PID_D);
-                            fy = ((_target_velocity.Y) - vel.Y) * (PID_D);
+                            fx = (float)((_target_velocity.X) - vel.X) * (PID_D);
+                            fy = (float)((_target_velocity.Y) - vel.Y) * (PID_D);
 
                             // vec.Z = (_target_velocity.Z - vel.Z) * PID_D + (_zeroPosition.Z - pos.Z) * PID_P;
 
-                            fz = fz + ((_target_velocity.Z - vel.Z) * (PID_D) * m_mass);
+                            fz = (float)(fz + ((_target_velocity.Z - vel.Z) * (PID_D) * m_mass));
                         }
                     }        // end if (m_usePID)
                     #endregion
@@ -1733,11 +1733,11 @@ Console.WriteLine(" JointCreateFixed");
                         switch (m_PIDHoverType)
                         {
                             case PIDHoverType.Ground:
-                                m_groundHeight = _parent_scene.GetTerrainHeightAtXY(pos.X, pos.Y);
+                                m_groundHeight = _parent_scene.GetTerrainHeightAtXY((float)pos.X, (float)pos.Y);
                                 m_targetHoverHeight = m_groundHeight + m_PIDHoverHeight;
                                 break;
                             case PIDHoverType.GroundAndWater:
-                                m_groundHeight = _parent_scene.GetTerrainHeightAtXY(pos.X, pos.Y);
+                                m_groundHeight = _parent_scene.GetTerrainHeightAtXY((float)pos.X, (float)pos.Y);
                                 m_waterHeight = _parent_scene.GetWaterLevel();
                                 if (m_groundHeight > m_waterHeight)
                                 {
@@ -1754,7 +1754,7 @@ Console.WriteLine(" JointCreateFixed");
 
                         _target_velocity =
                             new Vector3(0.0f, 0.0f,
-                                (m_targetHoverHeight - pos.Z) * ((PID_G - m_PIDHoverTau) * timestep)
+                                (float)(m_targetHoverHeight - pos.Z) * ((PID_G - m_PIDHoverTau) * timestep)
                                 );
 
                         //  if velocity is zero, use position control; otherwise, velocity control
@@ -1778,7 +1778,7 @@ Console.WriteLine(" JointCreateFixed");
                             _zeroFlag = false;
 
                             // We're flying and colliding with something
-                            fz = fz + ((_target_velocity.Z - vel.Z) * (PID_D) * m_mass);
+                            fz = (float)(fz + ((_target_velocity.Z - vel.Z) * (PID_D) * m_mass));
                         }
                     }
                     #endregion
@@ -1833,7 +1833,7 @@ Console.WriteLine(" JointCreateFixed");
                             {
                                 if (Math.Abs(vel.X) < 0.0001 || Math.Abs(vel.Y) < 0.0001 || Math.Abs(vel.Z) < 0.0001)
                                 {
-                                    Vector3 angvelocity = new Vector3(angvel.X, angvel.Y, angvel.Z);
+                                    Vector3 angvelocity = new Vector3((float)angvel.X, (float)angvel.Y, (float)angvel.Z);
                                     if (Math.Abs(force.X) < 100 && Math.Abs(force.Y) < 100 && Math.Abs(force.Z) < 100 &&
                                         angvelocity.ApproxEquals(Vector3.Zero, 0.01f) &&
                                         vel.X != 0 && vel.Y != 0 && vel.Z != 0)
@@ -2750,7 +2750,7 @@ Console.WriteLine(" JointCreateFixed");
                     d.Vector3 vel = d.BodyGetLinearVel(Body);
                     d.Vector3 rotvel = d.BodyGetAngularVel(Body);
                     d.Vector3 torque = d.BodyGetTorque(Body);
-                    _torque = new Vector3(torque.X, torque.Y, torque.Z);
+                    _torque = new Vector3((float)torque.X, (float)torque.Y, (float)torque.Z);
                     Vector3 l_position = Vector3.Zero;
                     Quaternion l_orientation = Quaternion.Identity;
 
@@ -2763,13 +2763,13 @@ Console.WriteLine(" JointCreateFixed");
                     m_lastposition = _position;
                     m_lastorientation = _orientation;
 
-                    l_position.X = vec.X;
-                    l_position.Y = vec.Y;
-                    l_position.Z = vec.Z;
-                    l_orientation.X = ori.X;
-                    l_orientation.Y = ori.Y;
-                    l_orientation.Z = ori.Z;
-                    l_orientation.W = ori.W;
+                    l_position.X = (float)vec.X;
+                    l_position.Y = (float)vec.Y;
+                    l_position.Z = (float)vec.Z;
+                    l_orientation.X = (float)ori.X;
+                    l_orientation.Y = (float)ori.Y;
+                    l_orientation.Z = (float)ori.Z;
+                    l_orientation.W = (float)ori.W;
 
                     if (l_position.X > ((int)_parent_scene.WorldExtents.X - 0.05f) || l_position.X < 0f || l_position.Y > ((int)_parent_scene.WorldExtents.Y - 0.05f) || l_position.Y < 0f)
                     {
@@ -2877,9 +2877,9 @@ Console.WriteLine(" JointCreateFixed");
                         {
                             _position = l_position;
 
-                            _velocity.X = vel.X;
-                            _velocity.Y = vel.Y;
-                            _velocity.Z = vel.Z;
+                            _velocity.X = (float)vel.X;
+                            _velocity.Y = (float)vel.Y;
+                            _velocity.Z = (float)vel.Z;
 
                             _acceleration = ((_velocity - m_lastVelocity) / 0.1f);
                             //m_log.Info("[PHYSICS]: V1: " + _velocity + " V2: " + m_lastVelocity + " Acceleration: " + _acceleration.ToString());
@@ -2887,7 +2887,7 @@ Console.WriteLine(" JointCreateFixed");
                             if (_velocity.ApproxEquals(pv, 0.5f))
                                 m_rotationalVelocity = pv;
                             else
-                                m_rotationalVelocity = new Vector3(rotvel.X, rotvel.Y, rotvel.Z);
+                                m_rotationalVelocity = new Vector3((float)rotvel.X, (float)rotvel.Y, (float)rotvel.Z);
 
                             if (_velocity.ApproxEquals(pv, 0.1f))
                             {
@@ -2918,10 +2918,10 @@ Console.WriteLine(" JointCreateFixed");
 
                             //m_log.Debug("ODE: " + m_rotationalVelocity.ToString());
                         }
-                        _orientation.X = ori.X;
-                        _orientation.Y = ori.Y;
-                        _orientation.Z = ori.Z;
-                        _orientation.W = ori.W;
+                        _orientation.X = (float)ori.X;
+                        _orientation.Y = (float)ori.Y;
+                        _orientation.Z = (float)ori.Z;
+                        _orientation.W = (float)ori.W;
                         if (!m_throttleUpdates || throttleCounter > _parent_scene.geomUpdatesPerThrottledUpdate)
                         {
                             if (_parent == null)
@@ -3126,17 +3126,17 @@ Console.WriteLine(" JointCreateFixed");
         public Matrix4 FromDMass(d.Mass pMass)
         {
             Matrix4 obj;
-            obj.M11 = pMass.I.M00;
-            obj.M12 = pMass.I.M01;
-            obj.M13 = pMass.I.M02;
+            obj.M11 = (float)pMass.I.M00;
+            obj.M12 = (float)pMass.I.M01;
+            obj.M13 = (float)pMass.I.M02;
             obj.M14 = 0;
-            obj.M21 = pMass.I.M10;
-            obj.M22 = pMass.I.M11;
-            obj.M23 = pMass.I.M12;
+            obj.M21 = (float)pMass.I.M10;
+            obj.M22 = (float)pMass.I.M11;
+            obj.M23 = (float)pMass.I.M12;
             obj.M24 = 0;
-            obj.M31 = pMass.I.M20;
-            obj.M32 = pMass.I.M21;
-            obj.M33 = pMass.I.M22;
+            obj.M31 = (float)pMass.I.M20;
+            obj.M32 = (float)pMass.I.M21;
+            obj.M33 = (float)pMass.I.M22;
             obj.M34 = 0;
             obj.M41 = 0;
             obj.M42 = 0;
